Global industry classification standard (GICS)

The Global Industry Classification Standard (GICS) is a standardized classification system for equities developed jointly by Morgan Stanley Capital International (MSCI) and Standard & Poor's. The GICS methodology is used by the MSCI indexes, which include domestic and international stocks, as well as by a large portion of the professional investment management community. The GICS hierarchy begins with 11 sectors and is followed by 24 industry groups, 68 industries, and 157 sub-industries. Each stock that is classified will have a coding at all four of these levels.

standards

Format

A data.table. Columns include:

sector id: integer. GICS sector ID.

sector name: integer. Correspondig GICS sector name.

industry group id: integer. GICS industry group ID.

industry group name: integer. Correspondig GICS industry group name.

industry id: integer. GICS industry ID.

industry name: integer. Correspondig GICS industry name.

subindustry id: integer. GICS subindustry ID.

subindustry name: integer. Correspondig GICS subindustry name.

Source

https://www.msci.com

Details

The Global Industry Classification Standard (GICS) system was established in 1999 and has since become widely followed through direct use by portfolio managers and benchmarking to MSCI indexes. All told, more than 26,000 stocks worldwide have been classified by GICS, accounting for more than 95 MSCI estimates that more than $3 trillion in assets is benchmarked to its MSCI funds, many of which are sector-specific.The main goal of GICS is to allow all market participants to classify stocks by standardized industry definitions. GICS is used to make portfolio diversification and overall asset allocation decisions from within a common framework.GICS competes with the Industry Classification Benchmark (ICB) system, which is maintained by Dow Jones and London's FTSE Group. In practice, most of the same sector and industry designations exist in both standards. More recently, the usefulness of GICS classifications has been called into question. Many of our current economic measurements are still products of the Industrial Age: when the companies that were growing and shaping the world were giants with huge physical plants and plenty of material products — companies like Exxon Mobile and GE. Tech companies as we know them today hadn't even thought of it.Times have changed, but we still treat business entities as if they come from the traditional industry boundaries. Industry walls are disintegrating at a rapid pace. Over the past five years, the likes of Apple and Google have made significant moves into the automotive, healthcare, media, and smart home markets, among many others. GICS may classify them as part of the Information Technology sector, but we know their reach extends far beyond.Critics contend its time we move from a vertical industry emphasis to one centered on business models instead. By updating the Global Industry Classification Standard to reflect the wider view taken by today’s business winners - new measures and standards can help investors, customers and employees manage new strategic landscapes with greater insights.

See also